+++ src/main/java/modelo/EcualizarImagen.java
@@ -6,7 +6,7 @@ public interface EcualizarImagen{
 
     BufferedImage Ecualizar(BufferedImage imagen, double[] datos);
     
-    void setDatos(int f1, int f2, int a, int pot);
+    void setDatos(int f1, int f2, double a, double pot);
 
     int[] getData();
 }
